+++ src/payments/stripe.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 import stripe
 from typing import Dict, Any, Optional, List
 
@@ -16,7 +18,7 @@ class StripePaymentService(PaymentServiceInterface):
     async def create_payment_intent(
         self,
         order: OrderModel,
-        amount: float,
+        amount: Decimal,
         currency: str = "usd"
     ) -> Dict[str, Any]:
         try:
@@ -77,7 +79,7 @@ class StripePaymentService(PaymentServiceInterface):
     async def process_refund(
         self,
         payment: PaymentModel,
-        amount: Optional[float] = None,
+        amount: Optional[Decimal] = None,
         reason: Optional[str] = None
     ) -> Dict[str, Any]:
         try:
